Switch to wrapping ember-cli-coffeescript
This seems like a cleaner overall approach. Rather than relying on Ember-CLI to order the execution of these filters, we can just leverage Broccoli itself, since this is exactly what it is designed to do.

With this commit, ember-cli-coffees6 now directly imports ember-cli-coffeescript, and invokes it internally on the output of the ES6 filter.